Refactor day 7: replace brute force with a binary search
While researching part 1, I tried the geometric mean and that was way
off. Then I tried other things and the median gave me the right answer,
but at that point I had so much doubt in my "stats foo", that I went
with the brute force approach.  It was to be expected to be of bad
performance but I wanted to see part 2 and make it work.  Now that I am
pretty confident, that median and geometric mean would be correct, I
still stick with the "search" approach.  This change makes the tests for
the final results faster by two magnitudes and I call it "good enough".