Day 15, part 2 (FAILED)
The final test for part-2 is off; using data/values from others the
value seems slightly to high.  Since the problem is in the biggest
dataset it is also most difficult to debug and I think it would need a
solution from another source to export the distance-map and compare the
results.  But for now I call it quits.